requirements.txt
asgiref==3.3.0
attrs==20.3.0
Django==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
drf-spectacular==0.11.0
gunicorn==20.0.4
importlib-metadata==2.0.0
inflection==0.5.1
jsonschema==3.2.0
psycopg2==2.8.6
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
whitenoise==5.2.0
zipp==3.4.0
